@@ -1,10 +1,10 @@
 class UserModel {
   final String uid;
-  final String name;
+  final String? name;
   final String email;
-  UserModel({this.uid, this.name, this.email});
+  UserModel({required this.uid, this.name, required this.email});
 
-  factory UserModel.fromMap(Map<String, dynamic> json) => new UserModel(
+  factory UserModel.fromMap(Map<String, dynamic> json) => UserModel(
         uid: json["uid"],
         email: json["email"],
       );
