@@ -1,20 +1,30 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_firebase/models/user.model.dart';
-import 'package:flutter_firebase/screens/authenticate/sign_in.dart';
-import 'package:flutter_firebase/screens/categories/list.dart';
-import 'package:flutter_firebase/screens/home/home.dart';
-import 'package:flutter_firebase/screens/wrapper.dart';
-import 'package:flutter_firebase/services/auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter_firebase_v2/authenticate/register.dart';
+import 'package:flutter_firebase_v2/authenticate/sign_in.dart';
+import 'package:flutter_firebase_v2/home.dart';
+import 'package:flutter_firebase_v2/models/user.model.dart';
+import 'package:flutter_firebase_v2/pages/categories/add_update.dart';
+import 'package:flutter_firebase_v2/pages/categories/list.dart';
+import 'package:flutter_firebase_v2/services/auth.dart';
 import 'package:provider/provider.dart';
+import 'firebase_options.dart';
 
-void main() {
-  runApp(MyApp());
+Future<void> main() async {
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return StreamProvider<UserModel>.value(
+      initialData: UserModel(uid: "", email: ""),
       value: AuthService().user,
       child: MaterialApp(
         title: 'Flutter - Connecting Firebase',
@@ -25,12 +35,16 @@ class MyApp extends StatelessWidget {
             buttonColor: Colors.blue[900],
           ),
         ),
-        home: Wrapper(),
+        home: FirebaseAuth.instance.currentUser != null
+            ? const HomePage()
+            : const SignInPage(),
         routes: <String, WidgetBuilder>{
-          CategoriesScreen.routeName: (BuildContext context) =>
-              CategoriesScreen(),
-          Home.routeName: (BuildContext context) => Home(),
-          SignIn.routeName: (BuildContext context) => SignIn(),
+          CategoriesAddUpdatePage.routeName: (context) =>
+              const CategoriesAddUpdatePage(),
+          CategoryListPage.routeName: (context) => const CategoryListPage(),
+          HomePage.routeName: (context) => const HomePage(),
+          RegisterPage.routeName: (context) => const RegisterPage(),
+          SignInPage.routeName: (context) => const SignInPage(),
         },
       ),
     );
