@@ -1,9 +1,10 @@
 bool validateEmail(String value) {
-  Pattern pattern =
+  String pattern =
       r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
-  RegExp regex = new RegExp(pattern);
-  if (!regex.hasMatch(value))
+  RegExp regex = RegExp(pattern);
+  if (!regex.hasMatch(value)) {
     return false;
-  else
+  } else {
     return true;
+  }
 }
